requirements.txt
Django==4.*
django_extensions==3.2.*
django-admin-rangefilter==0.8.*
django-import-export==2.8.*
djangorestframework==3.13.*
djangorestframework-simplejwt==5.2.*
django-filter==22.*
drf-excel==2.1.*
drf-yasg==1.21.*

ipython==8.4.*
Celery==5.2.*

#front
crispy-bootstrap5==0.6

#dev
django-debug-toolbar==3.5.*

#flake8
flake8==4.*
flake8-builtins==1.5.*
flake8-debugger==4.*
flake8-print==5.*

#pytest
pytest==7.1.*
pytest-django==4.5.*
pytest-cov==4.0.*
pytest-mock==3.9.*

#gunicorn
gunicorn==20.1.*

#uwsgi
uwsgi==2.0.*
